refactor(plugins): remove plugin utils under plugins dir
